requirements.txt
opencv-python

pillow
